fix: Reth default command (#400)
* refactor(charts): simplify reth node command execution in values.yaml

The `defaultCommandTemplate` in `charts/reth/values.yaml` has been refactored to simplify the execution of the `reth node` command. Previously, the command was split across multiple lines and arguments were appended separately.

The changes consolidate the command into a single line, directly embedding the `defaultCommandArgsTemplate` within the `exec` command. This improves readability and maintainability by ensuring the arguments are correctly passed to the `reth node` executable. The `trim` and `replace "\n" " "` functions are used to ensure that any leading/trailing whitespace or newlines in the arguments are handled correctly, preventing potential command execution errors.

* chore(charts): update chart version to 0.1.4